# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/accounts": {
      "get": {
        "operationId": "AccountsController_list",
        "summary": "List my trading accounts (contract §4.3)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TradingAccountDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AccountsController_connect",
        "summary": "Connect a trading account (contract §4.3)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradingAccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/accounts/{id}": {
      "delete": {
        "operationId": "AccountsController_remove",
        "summary": "Disconnect a trading account (contract §4.3)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/accounts/{id}/password": {
      "post": {
        "operationId": "AccountsController_changePassword",
        "summary": "Change an account password (contract §4.3)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAccountPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradingAccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Millimore API",
    "description": "Backend API for the Millimore platform (mobile app + admin dashboard). Single source of truth: docs/BACKEND_CONTRACT.md",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.millimore.app/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TradingAccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "acc_1"
          },
          "brokerId": {
            "type": "string",
            "example": "xm"
          },
          "brokerName": {
            "type": "string",
            "example": "XM"
          },
          "accountNumber": {
            "type": "string",
            "example": "50231487"
          },
          "masked": {
            "type": "string",
            "example": "••••1487"
          },
          "server": {
            "type": "string",
            "example": "XM-Live3"
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "balance": {
            "type": "number",
            "example": 5000
          },
          "status": {
            "type": "string",
            "enum": [
              "connected",
              "pending",
              "disconnected",
              "error"
            ]
          },
          "connectedAt": {
            "type": "string",
            "example": "2026-07-01T00:00:00Z"
          }
        },
        "required": [
          "id",
          "brokerId",
          "brokerName",
          "accountNumber",
          "masked",
          "server",
          "currency",
          "balance",
          "status",
          "connectedAt"
        ]
      },
      "ConnectAccountDto": {
        "type": "object",
        "properties": {
          "brokerId": {
            "type": "string",
            "example": "xm"
          },
          "accountNumber": {
            "type": "string",
            "example": "50231487"
          },
          "server": {
            "type": "string",
            "example": "XM-Live3"
          },
          "password": {
            "type": "string",
            "example": "investor-or-trade-password",
            "writeOnly": true
          },
          "currency": {
            "type": "string",
            "example": "USD"
          }
        },
        "required": [
          "brokerId",
          "accountNumber",
          "server",
          "password"
        ]
      },
      "ChangeAccountPasswordDto": {
        "type": "object",
        "properties": {
          "current": {
            "type": "string",
            "writeOnly": true
          },
          "next": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "current",
          "next"
        ]
      },
...
    }
  }
}